[WIP] vnc_db_proxy workaround for automation k8s scripts
Closes-Bug: 1789175
Depends-On: Ib4516ba278d9b2d006c9d58af45908ebbd2a9cf4

+++ tools/vncdbproxy.sh
@@ -1,18 +1,35 @@
 #!/bin/bash
 
-if [ "$1" == "" ]; then
-    echo "Usage: vncdbproxy.sh <CONFIGDBSERVER_IP>"
-    exit 1
-fi
-
 TOP=$(cd "$(dirname "$0")" && cd ../ && pwd)
 
-docker rm -f vncdbproxy
+Usage()
+{
+	echo "Usage: vncdbproxy.sh [-k] [-n <NetworkMode>] [CONFIGDBSERVER_IP]"
+}
 
+ServicesIpAddress='127.0.0.1'
+RemoveDockers=1
+Network='bridge' # This is default for `docker run` if no `--net` param is specified
+while :; do
+	case "$1" in
+		'-n') Network="$2"; shift 2;;
+		'-k') RemoveDockers=0; shift;;
+		'-h') Usage; exit 0;;
+		*) break;;
+	esac
+done
+
+if [ "$Network" = 'bridge' ]; then
+	[ "$1" = '' ] && { Usage; exit 1; }
+	ServicesIpAddress="$1"
+fi
+
+[ $RemoveDockers -eq 1 ] && docker rm -f vncdbproxy
 docker build "$TOP/docker/vnc_db_proxy/" -t vncdbproxy
 
 docker run \
     --name vncdbproxy \
+    --net "$Network" \
     -p 8091:8082 \
     -d \
     -e CONFIG_API_PORT=8082 \
@@ -21,7 +38,7 @@ docker run \
     -e log_local=true \
     -e AUTH_MODE=none \
     -e AAA_MODE=cloud-admin \
-    -e ZOOKEEPER_SERVERS="$1":2181 \
-    -e CONFIGDB_SERVERS="$1":9161 \
-    -e RABBITMQ_SERVERS="$1":5673 \
+    -e ZOOKEEPER_SERVERS="$ServicesIpAddress:2181" \
+    -e CONFIGDB_SERVERS="$ServicesIpAddress:9161" \
+    -e RABBITMQ_SERVERS="$ServicesIpAddress:5673" \
     vncdbproxy
